backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.25
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-socketio==5.10.0
eventlet==0.33.3
Werkzeug==3.0.1

